[Mqtt Handler] 펌웨어 다운로드 이벤트 파싱 오류 수정 (#120)
* fix/mqtt handler: MQTT 클라이언트 연결 안정성 강화 및 비동기 메시지 처리 적용

* fix/mqtt handler: 펌웨어 다운로드 이벤트 파싱 오류 수정

* fix/mqtt handler: 펌웨어 다운로드 이벤트 파싱 오류 수정

* refactor/mqtt-handler: 로그 출력 대상을 표준 출력으로 설정

* refactor/mqtt-handler: 로그 출력 방식을 fmt에서 log 로 변경

+++ mqtt-handler/mqttclient/subscriber.go
@@ -81,7 +81,7 @@ func parseDownloadCancelAck(msg mqtt.Message) (*types.FirmwareDownloadEvent, err
 }
 
 // 공통 이벤트 빌더 - 토픽에서 region/group/device ID 추출 + 이벤트 생성
-func buildDownloadEvent(topic, commandID, message, status string, progress int64, totalBytes int64, downloadBytes int64, speedKbps int64, verified bool, downloadTime int64) *types.FirmwareDownloadEvent {
+func buildDownloadEvent(topic, commandID, message, status string, progress int64, totalBytes int64, downloadBytes int64, speedKbps float64, verified bool, downloadTime float64) *types.FirmwareDownloadEvent {
 	topicParts := strings.Split(topic, "/")
 	deviceID, _ := strconv.ParseInt(topicParts[1], 10, 64)
